Dendrograms for evolutionary analysis.

The phylogram R package is a tool for for developing phylogenetic trees as deeply-nested lists known as "dendrogram" objects. It provides functions for conversion between "dendrogram" and "phylo" class objects, as well as several tools for command-line tree manipulation and import/export via Newick parenthetic text. This improves accessibility to the comprehensive range of object-specific analytical and tree-visualization functions found across a wide array of bioinformatic R packages.

Functions

A brief description of the primary phylogram functions are provided with links to their help pages below.

File import/export

read.dendrogram reads a Newick parenthetic text string from a file or text connection and creates an object of class "dendrogram"

write.dendrogram outputs an object of class "dendrogram" to a text string or file in Newick format

Object conversion

as.phylo.dendrogram converts a dendrogram to an object of class "phylo" "dendrogram"

as.dendrogram.phylo converts a "phylo" object to a dendrogram

Tree editing and manipulation

prune remove branches from a dendrogram object based on regular expression pattern matching

ladder reorders the branches of a dendrogram object to aid visualization

remidpoint recursively sets "midpoint" and "members" attributes for a nested list/dendrogram object

reposition shifts a dendrogram object up or down (or sideways if plotted horizontally)

as.cladogram modifies the "height" attributes of the nodes such that all leaves terminate at zero